+++ src/cpp_bindings/solver/gmres.cpp
@@ -31,6 +31,31 @@ void init_gmres(py::module_ &module_solver)
              py::arg("exec"), py::arg("system_matrix"), py::arg("max_iters"),
              py::arg("krylov_dim"), py::arg("tolerance"),
              py::arg("relative_stop_mode"))
+        .def(py::init([](std::shared_ptr<gko::Executor> exec,
+                         std::shared_ptr<const gko::LinOp> system_matrix,
+                         std::shared_ptr<const gko::LinOp> preconditioner,
+                         size_t max_iters, size_t krylov_dim,
+                         ValueType reduction_factor, bool relative_stop_mode) {
+                 auto stop_mode = (relative_stop_mode)
+                                      ? gko::stop::mode::rhs_norm
+                                      : gko::stop::mode::absolute;
+                 auto fact = gko::share(
+                     gko::solver::Gmres<ValueType>::build()
+                         .with_criteria(
+                             gko::stop::Iteration::build().with_max_iters(
+                                 max_iters),
+                             gko::stop::ResidualNorm<ValueType>::build()
+                                 .with_baseline(stop_mode)
+                                 .with_reduction_factor(reduction_factor))
+                         .with_krylov_dim(krylov_dim)
+                         .with_generated_preconditioner(preconditioner)
+                         .on(exec));
+                 return gko::share(fact->generate(system_matrix));
+             }),
+             py::arg("exec"), py::arg("system_matrix"),
+             py::arg("preconditioner"), py::arg("max_iters"),
+             py::arg("krylov_dim"), py::arg("tolerance"),
+             py::arg("relative_stop_mode"))
         .def("initialize_logger",
              [](gko::solver::Gmres<ValueType> &o) {
                  std::shared_ptr<gko::log::Convergence<ValueType>>