+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # Version 0.1.0 (unreleased)
+- Add ILU preconditioner bindings [#55](https://github.com/Helmholtz-AI-Energy/pyGinkgo/pull/55)
 - Aligning tests with code [#65](https://github.com/Helmholtz-AI-Energy/pyGinkgo/pull/65)
 - Add file config solver bindings [#64](https://github.com/Helmholtz-AI-Energy/pyGinkgo/pull/64)
 - Add GMRES solver bindings [#51](https://github.com/Helmholtz-AI-Energy/pyGinkgo/pull/51)